deleted test API for loading data, moved our dataloader into dataloader folder
