+++ server/pg.js
@@ -10,22 +10,23 @@ const pool = new Pool({
 
 pool.connect();
 
-const createFuncList = 
+const createMasterTable = 
   `CREATE TABLE IF NOT EXISTS 
    mastertable (id int UNIQUE NOT NULL, by varchar(100), text varchar DEFAULT null, descendants smallint DEFAULT null, type varchar, time int DEFAULT null, title varchar DEFAULT null,score smallint DEFAULT null, kids smallint DEFAULT null)`
 
 
-const testQuery = 
-`INSERT INTO mastertable (id, by, text, descendants, type, time, title, url, kids)
- VALUES (1,'testuser', 'hello', 3,'story',23789, 'hello world','wwww.google.com', 23)`
-
- const testQuery2 = 
-`INSERT INTO mastertable (id)
- VALUES (3)`
+const createCacheTable = 
+`CREATE TABLE IF NOT EXISTS
+ cached_results (id SERIAL PRIMARY KEY, query_string VARCHAR, labels VARCHAR, data VARCHAR)    `
 
 
+// const testQuery = 
+// `INSERT INTO mastertable (id, by, text, descendants, type, time, title, url, kids)
+//  VALUES (1,'testuser', 'hello', 3,'story',23789, 'hello world','wwww.google.com', 23)`
 
-pool.query( createFuncList, (err, res) => {
+
+
+pool.query( createMasterTable, (err, res) => {
   if (err) return err;
   return res
 });
